backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
pyjwt==2.8.0
bcrypt==4.1.2
Flask-Cors==4.0.0